+++ www/admin/include/header.php
@@ -374,9 +374,12 @@ if (strlen($scripts_hh) > 1) {
     echo "    <td height=20 align=left colspan=10>\n";
     echo "      <span class=\"font2 fgdefault\"> &nbsp; \n";
     echo "        <a href=\"$PHP_SELF?ADD=1000000\"> Show Scripts </a> &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; \n";
-    echo "        <a href=\"$PHP_SELF?ADD=1111111\"> Add A New Script</a> &nbsp; &nbsp; | &nbsp; &nbsp; \n";
-    echo "        <a href=\"$PHP_SELF?ADD=0email\"> Show Email Templates </a> &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; \n";
-    echo "        <a href=\"$PHP_SELF?ADD=1email\"> Add A New Email Template </a>\n";
+    echo "        <a href=\"$PHP_SELF?ADD=1111111\"> Add A New Script</a>\n";
+    if (!file_exists($WeBServeRRooT . '/admin/include/content/scripts/email_templates.php')) {
+        echo "         &nbsp; &nbsp; | &nbsp; &nbsp; \n";
+        echo "        <a href=\"$PHP_SELF?ADD=0email\"> Show Email Templates </a> &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; \n";
+        echo "        <a href=\"$PHP_SELF?ADD=1email\"> Add A New Email Template </a>\n";
+    }
     echo "      </span>\n";
     echo "    </td>\n";
     echo "  </tr>\n";
